tests/submission/PhilipTaylor/canvas/size.attributes.idl.set.zero.html

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>Canvas test: size.attributes.idl.set.zero</title>
<meta content="Philip Taylor" name="author">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/canvas-tests.js"></script>
<link href="../common/canvas-tests.css" rel="stylesheet">
</head><body class="show_output">

<h1><a href="index.size.html">size</a>.<a href="index.size.attributes.html">attributes</a>.<a href="index.size.attributes.idl.html">idl</a>.<a href="index.size.attributes.idl.set.html">set</a>.zero</h1>
<p class="desc">Setting width/height IDL attributes to 0</p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.size.width">size.width</a>
</li><li><a href="../annotated-spec/canvas.html#testrefs.size.height">size.height</a>
</li></ul>
</div>

<p class="output">Actual output:</p>
<canvas class="output" height="50" id="c" width="100"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
var t = async_test("Setting width/height IDL attributes to 0");
_addTest(function(canvas, ctx) {

canvas.width = 0;
canvas.height = 0;
_assertSame(canvas.width, 0, "canvas.width", "0");
_assertSame(canvas.height, 0, "canvas.height", "0");


});
</script>

<div id="log"></div>
</body></html>
